데미지 주고받기 , 사망 디버그, 사망시 게임오버 UI 임시

+++ Team_Project/Assets/01_Scripts/Player/BattleScripts/BattleStatus.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class BattleStatus : MonoBehaviour
 {
     // 최대체력
-    public float maxHp;
+    public int maxHp;
     // 현재 체력
     public float currentHp;
     // 공격력
